chore: Update page layout and components
This commit updates the page layout and components to improve the functionality and visual appeal of the website. The Participants component has been added to display a rotating list of participant types, providing more information about the target audience. The BentoGrid component now includes an embedded Google Maps iframe, replacing a previous section. The styling and layout of the BentoGrid, Footer, Sponsors, and Prizes components have been improved for consistency. Additionally, the Footer component now uses a different logo image. These changes enhance the user experience and overall quality of the website.