chore: Update npm dependencies to include react-intersection-observer
This commit updates the npm dependencies in the package.json and package-lock.json files to include the latest stable version of the react-intersection-observer library. This library is useful for implementing intersection-based triggers and lazy loading of components. By adding this dependency, the project can take advantage of the features provided by react-intersection-observer.